+++ src/wireless_communication_status_screen.c
@@ -48,7 +48,7 @@ static struct WirelessCommunicationStatusScreen * sStatusScreen;
 
 static void CB2_InitWirelessCommunicationScreen(void);
 static void Task_WirelessCommunicationScreen(u8);
-static void WCSS_AddTextPrinterParameterized(u8, u8, const u8 *, u8, u8, u8);
+static void WCSS_AddTextPrinterParameterized(u32, u8, const u8 *, u8, u8, u8);
 static bool32 UpdateCommunicationCounts(u32 *, u32 *, u32 *, u8);
 
 static const u16 sBgTiles_Pal[] = INCBIN_U16("graphics/interface/wireless_info_screen.gbapal");
@@ -311,7 +311,7 @@ static void Task_WirelessCommunicationScreen(u8 taskId)
 
 #undef tState
 
-static void WCSS_AddTextPrinterParameterized(u8 windowId, u8 fontId, const u8 * str, u8 x, u8 y, u8 mode)
+static void WCSS_AddTextPrinterParameterized(u32 windowId, u8 fontId, const u8 * str, u8 x, u8 y, u8 mode)
 {
     u8 color[3];
